#137: Move function find_latest_trusted_version from proofdb to crev-lib. This function reads from the proofdb but shouldn't be a function of the proofdb. At this point, the proofdb only depends on functionality lower in the dependency stack.
